fix: 修复sws_scale调用中潜在的类型转换警告
将dst_linesize数组的第一个元素显式转换为int类型，以匹配sws_scale函数参数类型，消除编译器关于符号不匹配的警告。

+++ core/player/MediaPlayer.cpp
@@ -1391,7 +1391,7 @@ void MediaPlayer::DecodeThread() {
                     QImage qimage(frame_data.width, frame_data.height, QImage::Format_ARGB32);
                     if (!qimage.isNull()) {
                         uint8_t* dst_slices[4] = {qimage.bits(), nullptr, nullptr, nullptr};
-                        int dst_linesize[4] = {qimage.bytesPerLine(), 0, 0, 0};
+                        int dst_linesize[4] = {static_cast<int>(qimage.bytesPerLine()), 0, 0, 0};
 
                         sws_scale(sws_ctx,
                                   frame_data.data,
